Rename katsdpcontim to katacomb
It contains Obit, "Software for the recently deceased".
