+++ katsdpcontim/Dockerfile.xenial.dev
@@ -143,11 +143,11 @@ ADD . /src/katsdppipelines
 RUN pip install -e /src/katsdppipelines/Obit-pipeline
 
 # Install katsdpcontim package
-ADD katsdpcontim/katsdpcontim /src/katsdpcontim
-RUN pip install -e /src/katsdpcontim
+ADD katsdpcontim/katacomb /src/katacomb
+RUN pip install -e /src/katacomb
 
 # Add katsdpcontim task configuration files
-ADD katsdpcontim/katsdpcontim/katsdpcontim/conf /obitconf
+ADD katsdpcontim/katacomb/katacomb/conf /obitconf
 
 # Add OBIT setup script
 ADD katsdpcontim/setup_obit.sh /bin/setup_obit.sh
