+++ katsdpcontim/katacomb/essays/configure_aips_disks.py
@@ -1,7 +1,7 @@
 import os
 from os.path import join as pjoin
 
-import katsdpcontim
+import katacomb
 
 AIPS_DIR = pjoin(os.sep, 'usr', 'local', 'AIPS')
