+++ katsdpcontim/katacomb/katacomb/conf/mfimage_nosc.in
