+++ katsdpcontim/katacomb/katacomb/obit_types.py
